Move plugins/webroot.py to _internal/
