+++ certbot/tests/renewal_test.py
@@ -31,7 +31,7 @@ class RenewalTest(test_util.ConfigTestCase):
     @mock.patch('certbot._internal.renewal.cli.set_by_cli')
     def test_webroot_params_conservation(self, mock_set_by_cli):
         # For more details about why this test is important, see:
-        # certbot.plugins.webroot_test::WebrootActionTest::test_webroot_map_partial_without_perform
+        # certbot._internal.plugins.webroot_test::WebrootActionTest::test_webroot_map_partial_without_perform
         from certbot._internal import renewal
         mock_set_by_cli.return_value = False
