--- docs/api/plugins/webroot.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.plugins.webroot`
-----------------------------------
-
-.. automodule:: certbot.plugins.webroot
-   :members:
